requirements.txt
PyJWT==2.10.1
requests==2.32.4
qrcode==8.2
Pillow==11.2.1